+++ partials/singles/articles/articles-header-basic-no-image.php
@@ -3,7 +3,7 @@
 ?>
 <div class="grid-row mb-4">
   <div class="grid-item is-s-24 offset-s-0 is-m-20 offset-m-2 is-l-18 offset-l-2 is-xl-16 offset-xl-4 is-xxl-16 offset-xxl-4">
-    <h1 id="single-articles-title" class="font-size-15 font-weight-bold mb-3"><?php the_title(); ?></h1>
+    <h1 id="single-articles-title" class="font-size-15 font-weight-bold mb-3" data-testid="post-title"><?php the_title(); ?></h1>
     <?php
       if (!empty($meta['_cmb_standfirst'])) {
     ?><h2 class="font-size-12 font-weight-bold mb-3 text-wrap-pretty"><?php echo $meta['_cmb_standfirst'][0]; ?></h2>